integration/requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.9.24
flake8==5.0.4
h11==0.14.0
idna==3.4
iniconfig==1.1.1
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-dotenv==0.21.0
selenium==4.4.3
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
wsproto==1.2.0
